% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicspace.R
\docType{package}
\name{basicspace}
\alias{basicspace-package}
\alias{basicspace}
\title{basicspace: Recovering a Basic Space from Issue Scales.}
\description{
Conducts Aldrich-McKelvey and Blackbox Scaling (Poole et al 2016)
\url{doi:10.18637/jss.v069.i07} to recover latent dimensions of judgment.
}
\section{basicspace functions}{

The main functions are \code{aldmck()} and \code{blackbox()} and \code{blackbox_transpose()}.
}

\author{
\strong{Maintainer}: Dave Armstrong \email{davearmstrong.ps@gmail.com}

Authors:
\itemize{
  \item Keith Poole \email{ktpoole@uga.edu}
  \item Howard Rosenthal \email{hr31@nyu.edu}
  \item Jeffrey Lewis \email{jblewis@ucla.edu}
  \item James Lo \email{lojames@usc.edu}
  \item Royce Carroll \email{rcarroll@rice.edu}
}

}
