% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotcdf}
\alias{plotcdf}
\title{Coordinate Cumulative Distribution Plot}
\usage{
plotcdf(x, ..., align = NULL, xlim = c(-2, 2))
}
\arguments{
\item{x}{an \code{aldmck} output object.}

\item{...}{other arguments to \code{plot}.}

\item{align}{integer, the x-axis location that stimuli names should be
aligned to If set to NULL, it will attempt to guess a location.}

\item{xlim}{vector of length 2, fed to the \code{plot} function as the
\code{xlim} argument, which sets the minimum and maximum range of the
x-axis.}
}
\value{
A plot of the empirical cumulative distribution of the respondent
ideal points, along with the locations of the stimuli. If no self-placements
were specified during estimation, no graphical plots will appear.
}
\description{
Plots the cumulative distribution of the respondents and stimuli.
}
\examples{

### Loads and scales the Liberal-Conservative scales from the 1980 NES.
data(LC1980)
result <- aldmck(data=LC1980, polarity=2, respondent=1, missing=c(0,8,9),verbose=TRUE)

summary(result)
par(ask=TRUE)
plotcdf(result)

}
\seealso{
'\link{aldmck}', '\link{LC1980}', '\link{summary.aldmck}',
'\link{plot.aldmck}'.
}
\keyword{multivariate}
