% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootbbt}
\alias{bootbbt}
\title{Blackbox Transpose Bootstrap of 1980 Liberal-Conservative Scales.}
\format{
\subsection{\code{bootbbt}}{

A data frame with 10 rows and 6 columns giving the bootstrap placements for Carter, Reagan, Kennedy, Anderson, Republicans and Democrats
}
}
\usage{
bootbbt
}
\description{
Output from 10 bootstrap trials of LC1980 data. Included to allow the
example to run sufficiently quickly to pass CRAN guidelines.
}
\keyword{datasets}
