% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colombia}
\alias{colombia}
\title{2004 PELA Liberal-Conservative Scales.}
\format{
\subsection{\code{colombia}}{

An integer matrix with 95 rows and 17 columns:
\describe{
\item{id}{Respondent ID.}
\item{party}{Respondent party.}
\item{departam}{Respondent district.}
\item{entrev}{Interviewer ID.}
\item{pl_uribista}{Placement of Partido Liberal Uribista on 10 point scale.}
\item{pl_oficial}{Placement of Partido Liberal Oficial on 10 point scale.}
\item{conservator}{Placement of Partido Conservador on 10 point scale.}
\item{polo}{Placement of Polo on 10 point scale.}
\item{union_cristiana}{Placement of Union Cristiana on 10 point scale.}
\item{salvation}{Placement of Salvacion on 10 point scale.}
\item{urine}{Placement of Mr. Uribe on 10 point scale.}
\item{antanas}{Placement of Mr. Antanas on 10 point scale.}
\item{gomez}{Placement of Mr. Gomez on 10 point scale.}
\item{garzon}{Placement of Garzon on 10 point scale.}
\item{holgin}{Placement of Holguin on 10 point scale.}
\item{rivera}{Placement of Rivera on 10 point scale.}
\item{self}{Respondent self placement on 10 point scale.}
}
}
}
\source{
Sebastian Saiegh. 2009. Recovering a Basic Space from Elite Surveys:
Evidence from Latin America. Legislative Studies Quarterly. 34(1): 117-145. \doi{10.3162/036298009787500349}

Sebastian Saiegh. 2011. Ruling By Statute: How Uncertainty and Vote-Buying
Shape Lawmaking. New York: Cambridge University Press. \doi{10.1017/CBO9780511842276}
}
\usage{
colombia
}
\description{
Liberal-Conservative 10-point scales from the University of Salamanca's
Parliamentary Elites of Latin America (PELA) survey. Stored as a matrix of
integers. The number 99 is a missing value. These data come from Sebastian
Saiegh and are used in the paper and book cited below.
}
\keyword{datasets}
