% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Issues1980}
\alias{Issues1980}
\title{1980 Issues Scales}
\format{
\subsection{\code{Issues1980}}{

A matrix, containing reported self-placements along various stimuli on a  7 point
Liberal-Conservative scales (with the exception of abortion scales, which are 4 point):
\describe{
\item{libcon1}{Liberal-conservative self-placement on 7 point scale.}
\item{defense}{Defense spending self-placement on 7 point scale.}
\item{govserv}{Government service on 7 point scale.}
\item{inflation}{Importance of inflation self-placement on 7 point scale.}
\item{abortion1}{Attitude on abortion 4 point scale.}
\item{taxcut}{Support for tax cut on 7 point scale.}
\item{libcon2}{Liberal-conservative self-placement on 7 point scale.}
\item{govhelpmin}{Government aid on 7 point scale.}
\item{russia}{Attitude towards Russia on 7 point scale.}
\item{womenrole}{Role of women on 7 point scale.}
\item{govjobs}{Placement of Democrats on 7 point scale.}
\item{equalrights}{Support for equal rights on 7 point scale.}
\item{busing}{Opinion on busing on 7 point scale.}
\item{abortion2}{Another attitude on abortion on 4 point scale.}
}
}
}
\source{
American National Election Study. \url{https://electionstudies.org}
Also available from Keith Poole's website. \url{https://voteview.com}
}
\usage{
Issues1980
}
\description{
Issue scales from the 1980 National Election Study. The numbers 0, 8, and 9
are considered to be missing values, except for the two abortion scales,
where '7' is also a missing value.  Hence, it must be recoded as in the
example shown below before scaling. The data is used as an example for
blackbox().
}
\keyword{datasets}
