% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Issues1980_bb}
\alias{Issues1980_bb}
\title{Blackbox Estimate, 1980 NES Issue Scales.}
\format{
\subsection{\code{Issues1980_bb}}{

An object of class \code{blackbox} with the following elements:
\describe{
\item{stimuli}{A list of length 3 (number of dimensions in the scaling analysis) where each element contains a data frame with the following variables:}
\itemize{
\item {N}Number of respondents who provided a response to this stimulus.
\item{c}Stimulus intercept.
\item{w1}Estimate of the stimulus weight on the first dimension. If viewing the results for a higher dimension, higher dimension results will appear as w2, w3, etc.
\item{R2}The percent variance explained for the stimulus. This increases as more dimensions are estimated.
}
\item{individuals}{A list of length 3 (number of dimensions in the scaling analysis) where each element is a data frame presenting results from the analysis fot he corresponding dimension.  The datasets contain the following variables:}
\itemize{
\item{c1}Estimate of the individual intercept on the first dimension. If viewing the results for a higher dimension, higher dimension results will appear as c2, c3, etc.
}
\item{fits}{ A data frame of fit results, with elements listed as follows:}
\itemize{
\item{SSE}Sum of squared errors.
\item{SSE.explained}Explained sum of squared error.
\item{percent}Percentage of total variance explained.
\item{SE}Standard error of the estimate, with formula provided on pg. 973 of the article cited below.
\item{singular}Singluar value for the dimension.
}
\item{Nrow}{Numberof rows/stimuli.}
\item{Ncol}{Number of columns used in estimation. This may differ from the data set due to columns discarded due to the \code{minscale} constraint.}
\item{Ndata}{Total number of data entries.}
\item{Nmiss}{Number of missing entries.}
\item{SS_mean}{ Sum of squares grand mean.}
\item{dims}{ Number of dimensions estimated.}
}
}
}
\source{
American National Election Study. \url{https://electionstudies.org}
}
\usage{
data(Issues1980_bb)
}
\description{
Blackbox estimates from issues scales from the 1980 National Election Study.
}
\keyword{datasets}
