% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LC1980}
\alias{LC1980}
\title{1980 Liberal-Conservative Scales.}
\format{
\subsection{\code{LC1980}}{

matrix, containing reported placements of various stimuli on a 7 point Liberal-Conservative scale:
\describe{
\item{Self}{Self-placement on 7 point scale.}
\item{Carter}{Placement of Carter on 7 point scale.}
\item{Reagan}{Placement of Reagan on 7 point scale.}
\item{Kennedy}{Placement of Kennedy on 7 point scale.}
\item{Anderson}{Placement of Anderson on 7 point scale.}
\item{Republicans}{Placement of Republicans on 7 point scale.}
\item{Democrats}{Placement of Democrats on 7 point scale.}
}
}
}
\source{
American National Election Study. \url{https://electionstudies.org}
Also available from Keith Poole's website. \url{https://voteview.com}
}
\usage{
LC1980
}
\description{
Liberal-Conservative 7-point scales from the 1980 National Election Study.
Includes (in order) self-placement, and rankings of Carter, Reagan, Kennedy,
Anderson, Republican party, Democratic Party. Stored as a matrix of
integers. The numbers 0, 8, and 9 are considered to be missing values.
}
\keyword{datasets}
