% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackbox.R
\name{plot.blackbox}
\alias{plot.blackbox}
\title{Blackbox Coordinate Distribution Plot}
\usage{
\method{plot}{blackbox}(x, ...)
}
\arguments{
\item{x}{an \code{blackbox} output object.}

\item{...}{other arguments to \code{hist}.}
}
\value{
A histogram of the estimated intercepts.
}
\description{
\code{plot.blackbox} reads an \code{blackbox} object and plots a histogram
of the estimated intercepts.
}
\examples{



### Loads issue scales from the 1980 NES.
data(Issues1980)
Issues1980[Issues1980[,"abortion1"]==7,"abortion1"] <- 8	#missing recode
Issues1980[Issues1980[,"abortion2"]==7,"abortion2"] <- 8	#missing recode

### This command conducts estimates, which we instead load using data()
# Issues1980_bb <- blackbox(Issues1980,missing=c(0,8,9),verbose=FALSE,dims=3,minscale=8)
data(Issues1980_bb)

plot(Issues1980_bb)


}
\seealso{
\link{Issues1980}.
}
\keyword{multivariate}
