% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldmck.R
\name{plot.boot_aldmck}
\alias{plot.boot_aldmck}
\title{Bootstrapped Aldrich-McKelvey Stimulus Plots}
\usage{
\method{plot}{boot_aldmck}(x, ...)
}
\arguments{
\item{x}{an \code{boot_aldmck} output object.}

\item{...}{other arguments to \code{plot}.}
}
\value{
A dotchart of estimated stimulus locations, with 95 percent confidence
intervals. Point estimates are estimates from the original data set.
}
\description{
\code{plot.boot_aldmck} reads an \code{boot_aldmck} object and plots a
dotchart of the stimuli with estimated confidence intervals.
}
\examples{



data(LC1980)
result <- boot_aldmck(data=LC1980, polarity=2, respondent=1, missing=c(0,8,9), iter=30)
plot(result)


}
\seealso{
\link{aldmck}, \link{boot_aldmck}.
}
\keyword{multivariate}
