% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stimuli}
\alias{stimuli}
\title{Stimulus extraction function}
\usage{
stimuli(object)
}
\arguments{
\item{object}{an \code{aldmck}, \code{blackbox}, or \code{blackbt} output
object.}
}
\value{
The stimuli of the estimated output, which can also be recovered as
\code{object$stimuli}.  Please refer to the documentation of \code{aldmck},
\code{blackbox}, or \code{blackbox_transpose} for specifics.
}
\description{
\code{stimuli} is a convenience function to extract the stimulus parameters
from an \code{aldmck}, \code{blackbox}, or \code{blackbt} object.
}
\examples{

data(Issues1980)
Issues1980[Issues1980[,"abortion1"]==7,"abortion1"] <- 8	#missing recode
Issues1980[Issues1980[,"abortion2"]==7,"abortion2"] <- 8	#missing recode

### This command conducts estimates, which we instead load using data()
# Issues1980_bb <- blackbox(Issues1980,missing=c(0,8,9),verbose=FALSE,dims=3,minscale=8)
data(Issues1980_bb)

stimuli(Issues1980_bb)

}
\seealso{
'\link{aldmck}', '\link{blackbox}', '\link{blackbox_transpose}'.
}
\keyword{multivariate}
